+++ great_expectations/cli/datasource.py
@@ -61,6 +61,7 @@ class SupportedDatabases(enum.Enum):
     POSTGRES = 'Postgres'
     REDSHIFT = 'Redshift'
     SNOWFLAKE = 'Snowflake'
+    BIGQUERY = 'BigQuery'
     OTHER = 'other - Do you have a working SQLAlchemy connection string?'
     # TODO MSSQL
     # TODO BigQuery
@@ -433,6 +434,10 @@ def _add_sqlalchemy_datasource(context, prompt_for_datasource_name=True):
             if not load_library("snowflake", install_instructions_string="pip install snowflake-sqlalchemy"):
                 return None
             credentials = _collect_snowflake_credentials(default_credentials=credentials)
+        elif selected_database == SupportedDatabases.BIGQUERY:
+            if not load_library("pybigquery", install_instructions_string="pip install pybigquery"):
+                return None
+            credentials = _collect_bigquery_credentials(default_credentials=credentials)
         elif selected_database == SupportedDatabases.OTHER:
             sqlalchemy_url = click.prompt(
                 """What is the url/connection string for the sqlalchemy connection?
@@ -573,6 +578,17 @@ def _collect_snowflake_credentials(default_credentials=None):
 
     return credentials
 
+def _collect_bigquery_credentials(default_credentials=None):
+    sqlalchemy_url = click.prompt(
+"""What is the SQLAlchemy url/connection string for the BigQuery connection?
+(reference: https://github.com/mxmzdlv/pybigquery#connection-string-parameters)
+""",
+        show_default=False).strip()
+    credentials = {
+        "url": sqlalchemy_url
+    }
+
+    return credentials
 
 def _collect_mysql_credentials(default_credentials=None):
     # We are insisting on pymysql driver when adding a MySQL datasource through the CLI